requirements.txt
cupy-cuda113==10.6.0
ffmpeg-python==0.2.0
imageio==2.19.3
imageio-ffmpeg==0.4.7
Markdown==3.4.1
numpy==1.23.5
opencv-python-headless==4.6.0.66
protobuf==3.20.2
sk-video==1.1.10
ffmpy==0.3.0
gradio==3.36.1
matplotlib==3.6.3
PyYAML==6.0
timm==0.6.12
tqdm==4.64.1
pexpect==4.8.0